+++ src/pop/integrations/langchain/__init__.py
@@ -2,10 +2,18 @@ from pop.integrations.langchain.bindings import (
     bind_langchain_context,
     bind_langchain_middleware,
     bind_langchain_prompt,
+    build_langchain_agent_input,
+    create_langchain_agent_kwargs,
+    create_langchain_execution_scaffold,
+    create_langchain_middleware_scaffold,
 )
 
 __all__ = [
     "bind_langchain_prompt",
     "bind_langchain_context",
     "bind_langchain_middleware",
+    "build_langchain_agent_input",
+    "create_langchain_agent_kwargs",
+    "create_langchain_execution_scaffold",
+    "create_langchain_middleware_scaffold",
 ]