+++ src/pop_protocol/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.1.2"
+__version__ = "0.1.3"
 
 from pop_protocol.core import load_pop